test(scan): consolidate scanner-agnostic tests
This recombines the scanner-agnostic tests in the duplicated `app_config.test.ts` files into a single `app_config.test.ts`. It also consolidates some of the `createApp` and related helper functionality.
