+++ apps/scan/backend/src/scanners/plustek/app_scan.test.ts
@@ -6,12 +6,14 @@ import { MAX_FAILED_SCAN_ATTEMPTS } from './state_machine';
 import { PrecinctScannerInterpreter } from '../../interpret';
 import {
   ballotImages,
-  configureApp,
-  createApp,
-  expectStatus,
-  waitForStatus,
+  createPlustekScannerApp,
 } from '../../../test/helpers/scanners/plustek/app_helpers';
 import { SheetInterpretation } from '../../types';
+import {
+  configureApp,
+  expectStatus,
+  waitForStatus,
+} from '../../../test/helpers/app_helpers';
 
 jest.setTimeout(20_000);
 jest.mock('@votingworks/ballot-encoder', () => {
@@ -94,7 +96,8 @@ function mockInterpretation(
 }
 
 test('configure and scan hmpb', async () => {
-  const { apiClient, mockPlustek, mockUsb, logger } = await createApp();
+  const { apiClient, mockPlustek, mockUsb, logger } =
+    await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb, { addTemplates: true });
 
   (
@@ -133,7 +136,8 @@ test('configure and scan hmpb', async () => {
 });
 
 test('configure and scan bmd ballot', async () => {
-  const { apiClient, mockPlustek, mockUsb, logger } = await createApp();
+  const { apiClient, mockPlustek, mockUsb, logger } =
+    await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb);
 
   (
@@ -180,7 +184,7 @@ const needsReviewInterpretation: SheetInterpretation = {
 
 test('ballot needs review - return', async () => {
   const { apiClient, mockPlustek, workspace, mockUsb, logger } =
-    await createApp();
+    await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb, { addTemplates: true });
 
   (
@@ -220,7 +224,8 @@ test('ballot needs review - return', async () => {
 });
 
 test('ballot needs review - accept', async () => {
-  const { apiClient, mockPlustek, mockUsb, logger } = await createApp();
+  const { apiClient, mockPlustek, mockUsb, logger } =
+    await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb, { addTemplates: true });
 
   (
@@ -260,7 +265,7 @@ test('ballot needs review - accept', async () => {
 // TODO test all the invalid ballot reasons?
 test('invalid ballot rejected', async () => {
   const { apiClient, mockPlustek, workspace, mockUsb, logger } =
-    await createApp();
+    await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb);
 
   (
@@ -298,7 +303,7 @@ test('invalid ballot rejected', async () => {
 });
 
 test('bmd ballot is rejected when scanned for wrong precinct', async () => {
-  const { apiClient, mockPlustek, mockUsb } = await createApp();
+  const { apiClient, mockPlustek, mockUsb } = await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb, { precinctId: '22' });
   // Ballot should be rejected when configured for the wrong precinct
 
@@ -328,7 +333,7 @@ test('bmd ballot is rejected when scanned for wrong precinct', async () => {
 });
 
 test('bmd ballot is accepted if precinct is set for the right precinct', async () => {
-  const { apiClient, mockPlustek, mockUsb } = await createApp();
+  const { apiClient, mockPlustek, mockUsb } = await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb, { precinctId: '23' });
   // Configure for the proper precinct and verify the ballot scans
 
@@ -350,7 +355,7 @@ test('bmd ballot is accepted if precinct is set for the right precinct', async (
 });
 
 test('hmpb ballot is rejected when scanned for wrong precinct', async () => {
-  const { apiClient, mockPlustek, mockUsb } = await createApp();
+  const { apiClient, mockPlustek, mockUsb } = await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb, {
     addTemplates: true,
     precinctId: '22',
@@ -383,7 +388,7 @@ test('hmpb ballot is rejected when scanned for wrong precinct', async () => {
 });
 
 test('hmpb ballot is accepted if precinct is set for the right precinct', async () => {
-  const { apiClient, mockPlustek, mockUsb } = await createApp();
+  const { apiClient, mockPlustek, mockUsb } = await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb, {
     addTemplates: true,
     precinctId: '21',
@@ -408,7 +413,7 @@ test('hmpb ballot is accepted if precinct is set for the right precinct', async 
 });
 
 test('blank sheet ballot rejected', async () => {
-  const { apiClient, mockPlustek, mockUsb } = await createApp();
+  const { apiClient, mockPlustek, mockUsb } = await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb);
 
   (await mockPlustek.simulateLoadSheet(ballotImages.blankSheet)).unsafeUnwrap();
@@ -435,7 +440,7 @@ test('blank sheet ballot rejected', async () => {
 });
 
 test('scanner powered off while waiting for paper', async () => {
-  const { apiClient, mockPlustek, mockUsb } = await createApp();
+  const { apiClient, mockPlustek, mockUsb } = await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb);
 
   mockPlustek.simulatePowerOff();
@@ -446,7 +451,7 @@ test('scanner powered off while waiting for paper', async () => {
 });
 
 test('scanner powered off while scanning', async () => {
-  const { apiClient, mockPlustek, mockUsb } = await createApp();
+  const { apiClient, mockPlustek, mockUsb } = await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb);
 
   (
@@ -464,7 +469,8 @@ test('scanner powered off while scanning', async () => {
 });
 
 test('scanner powered off while accepting', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb } = await createApp();
+  const { apiClient, mockPlustek, interpreter, mockUsb } =
+    await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb);
 
   (
@@ -496,7 +502,8 @@ test('scanner powered off while accepting', async () => {
 });
 
 test('scanner powered off after accepting', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb } = await createApp();
+  const { apiClient, mockPlustek, interpreter, mockUsb } =
+    await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb);
 
   (
@@ -534,7 +541,8 @@ test('scanner powered off after accepting', async () => {
 });
 
 test('scanner powered off while rejecting', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb } = await createApp();
+  const { apiClient, mockPlustek, interpreter, mockUsb } =
+    await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb);
 
   (
@@ -563,7 +571,8 @@ test('scanner powered off while rejecting', async () => {
 });
 
 test('scanner powered off while returning', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb } = await createApp();
+  const { apiClient, mockPlustek, interpreter, mockUsb } =
+    await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb);
 
   (
@@ -592,7 +601,8 @@ test('scanner powered off while returning', async () => {
 });
 
 test('scanner powered off after returning', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb } = await createApp();
+  const { apiClient, mockPlustek, interpreter, mockUsb } =
+    await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb);
 
   (
@@ -628,7 +638,7 @@ test('scanner powered off after returning', async () => {
 });
 
 test('insert second ballot while first ballot is scanning', async () => {
-  const { apiClient, mockPlustek, mockUsb } = await createApp({
+  const { apiClient, mockPlustek, mockUsb } = await createPlustekScannerApp({
     delays: {},
     mockPlustekOptions: { passthroughDuration: 500 },
   });
@@ -662,7 +672,8 @@ test('insert second ballot while first ballot is scanning', async () => {
 });
 
 test('insert second ballot before first ballot accept', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb } = await createApp();
+  const { apiClient, mockPlustek, interpreter, mockUsb } =
+    await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb);
 
   (
@@ -700,12 +711,13 @@ test('insert second ballot before first ballot accept', async () => {
 });
 
 test('insert second ballot while first ballot is accepting', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb } = await createApp({
-    delays: {
-      DELAY_ACCEPTED_READY_FOR_NEXT_BALLOT: 1000,
-      DELAY_ACCEPTED_RESET_TO_NO_PAPER: 2000,
-    },
-  });
+  const { apiClient, mockPlustek, interpreter, mockUsb } =
+    await createPlustekScannerApp({
+      delays: {
+        DELAY_ACCEPTED_READY_FOR_NEXT_BALLOT: 1000,
+        DELAY_ACCEPTED_RESET_TO_NO_PAPER: 2000,
+      },
+    });
   await configureApp(apiClient, mockUsb);
 
   (
@@ -738,7 +750,8 @@ test('insert second ballot while first ballot is accepting', async () => {
 });
 
 test('insert second ballot while first ballot needs review', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb } = await createApp();
+  const { apiClient, mockPlustek, interpreter, mockUsb } =
+    await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb);
 
   (
@@ -773,10 +786,11 @@ test('insert second ballot while first ballot needs review', async () => {
 });
 
 test('insert second ballot while first ballot is rejecting', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb } = await createApp({
-    delays: {},
-    mockPlustekOptions: { passthroughDuration: 500 },
-  });
+  const { apiClient, mockPlustek, interpreter, mockUsb } =
+    await createPlustekScannerApp({
+      delays: {},
+      mockPlustekOptions: { passthroughDuration: 500 },
+    });
   await configureApp(apiClient, mockUsb);
 
   (
@@ -820,10 +834,11 @@ test('insert second ballot while first ballot is rejecting', async () => {
 });
 
 test('insert second ballot while first ballot is returning', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb } = await createApp({
-    delays: {},
-    mockPlustekOptions: { passthroughDuration: 500 },
-  });
+  const { apiClient, mockPlustek, interpreter, mockUsb } =
+    await createPlustekScannerApp({
+      delays: {},
+      mockPlustekOptions: { passthroughDuration: 500 },
+    });
   await configureApp(apiClient, mockUsb);
 
   (
@@ -863,7 +878,7 @@ test('insert second ballot while first ballot is returning', async () => {
 });
 
 test('jam on scan', async () => {
-  const { apiClient, mockPlustek, mockUsb } = await createApp({
+  const { apiClient, mockPlustek, mockUsb } = await createPlustekScannerApp({
     delays: {
       DELAY_RECONNECT_ON_UNEXPECTED_ERROR: 500,
     },
@@ -885,11 +900,12 @@ test('jam on scan', async () => {
 });
 
 test('jam on accept', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb } = await createApp({
-    delays: {
-      DELAY_ACCEPTING_TIMEOUT: 500,
-    },
-  });
+  const { apiClient, mockPlustek, interpreter, mockUsb } =
+    await createPlustekScannerApp({
+      delays: {
+        DELAY_ACCEPTING_TIMEOUT: 500,
+      },
+    });
   await configureApp(apiClient, mockUsb);
 
   (
@@ -927,7 +943,8 @@ test('jam on accept', async () => {
 });
 
 test('jam on return', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb } = await createApp();
+  const { apiClient, mockPlustek, interpreter, mockUsb } =
+    await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb);
 
   (
@@ -954,7 +971,8 @@ test('jam on return', async () => {
 });
 
 test('jam on reject', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb } = await createApp();
+  const { apiClient, mockPlustek, interpreter, mockUsb } =
+    await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb);
 
   (
@@ -981,7 +999,7 @@ test('jam on reject', async () => {
 });
 
 test('calibrate', async () => {
-  const { apiClient, mockPlustek, mockUsb } = await createApp();
+  const { apiClient, mockPlustek, mockUsb } = await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb);
 
   (await mockPlustek.simulateLoadSheet(ballotImages.blankSheet)).unsafeUnwrap();
@@ -994,7 +1012,7 @@ test('calibrate', async () => {
 });
 
 test('calibrate not supported', async () => {
-  const { apiClient, mockPlustek, mockUsb } = await createApp();
+  const { apiClient, mockPlustek, mockUsb } = await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb);
 
   (await mockPlustek.simulateLoadSheet(ballotImages.blankSheet)).unsafeUnwrap();
@@ -1006,7 +1024,7 @@ test('calibrate not supported', async () => {
 });
 
 test('jam on calibrate', async () => {
-  const { apiClient, mockPlustek, mockUsb } = await createApp();
+  const { apiClient, mockPlustek, mockUsb } = await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb);
 
   (await mockPlustek.simulateLoadSheet(ballotImages.blankSheet)).unsafeUnwrap();
@@ -1019,7 +1037,7 @@ test('jam on calibrate', async () => {
 
 test('scan fails and retries', async () => {
   const { apiClient, mockPlustek, logger, interpreter, mockUsb } =
-    await createApp();
+    await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb);
 
   (
@@ -1053,7 +1071,7 @@ test('scan fails and retries', async () => {
 });
 
 test('scan fails repeatedly and eventually gives up', async () => {
-  const { apiClient, mockPlustek, mockUsb } = await createApp();
+  const { apiClient, mockPlustek, mockUsb } = await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb);
 
   (
@@ -1077,7 +1095,8 @@ test('scan fails repeatedly and eventually gives up', async () => {
 });
 
 test('scan fails due to plustek returning only one file instead of two', async () => {
-  const { apiClient, mockPlustek, mockUsb, logger } = await createApp();
+  const { apiClient, mockPlustek, mockUsb, logger } =
+    await createPlustekScannerApp();
   await configureApp(apiClient, mockUsb);
 
   (
@@ -1108,12 +1127,13 @@ test('scan fails due to plustek returning only one file instead of two', async (
 });
 
 test('scanning time out', async () => {
-  const { apiClient, mockPlustek, logger, mockUsb } = await createApp({
-    delays: {
-      DELAY_SCANNING_TIMEOUT: 50,
-      DELAY_RECONNECT_ON_UNEXPECTED_ERROR: 500,
-    },
-  });
+  const { apiClient, mockPlustek, logger, mockUsb } =
+    await createPlustekScannerApp({
+      delays: {
+        DELAY_SCANNING_TIMEOUT: 50,
+        DELAY_RECONNECT_ON_UNEXPECTED_ERROR: 500,
+      },
+    });
   await configureApp(apiClient, mockUsb);
 
   (
@@ -1144,7 +1164,7 @@ test('scanning time out', async () => {
 });
 
 test('kills plustekctl if it freezes', async () => {
-  const { apiClient, mockPlustek, mockUsb } = await createApp({
+  const { apiClient, mockPlustek, mockUsb } = await createPlustekScannerApp({
     delays: {
       DELAY_SCANNING_TIMEOUT: 50,
       DELAY_RECONNECT_ON_UNEXPECTED_ERROR: 500,
@@ -1171,7 +1191,7 @@ test('kills plustekctl if it freezes', async () => {
 });
 
 test('stops completely if plustekctl freezes and cant be killed', async () => {
-  const { apiClient, mockPlustek, mockUsb } = await createApp({
+  const { apiClient, mockPlustek, mockUsb } = await createPlustekScannerApp({
     delays: {
       DELAY_SCANNING_TIMEOUT: 50,
       DELAY_RECONNECT_ON_UNEXPECTED_ERROR: 500,
