+++ apps/scan/backend/test/helpers/scanners/custom/app_helpers.ts
@@ -1,142 +1,40 @@
 import {
-  ALL_PRECINCTS_SELECTION,
-  singlePrecinctSelectionFor,
-} from '@votingworks/utils';
+  buildMockInsertedSmartCardAuth,
+  InsertedSmartCardAuthApi,
+} from '@votingworks/auth';
+import { deferred, ok, Result } from '@votingworks/basics';
+import {
+  electionFamousNames2021Fixtures,
+  sampleBallotImages,
+} from '@votingworks/fixtures';
 import * as grout from '@votingworks/grout';
-import { Application } from 'express';
-import { Buffer } from 'buffer';
-import { ElectionDefinition, PrecinctId } from '@votingworks/types';
-import waitForExpect from 'wait-for-expect';
 import { fakeLogger, Logger } from '@votingworks/logging';
 import {
   MockScannerClient,
   MockScannerClientOptions,
   ScannerClient,
 } from '@votingworks/plustek-scanner';
+import { Application } from 'express';
 import tmp from 'tmp';
-import {
-  electionFamousNames2021Fixtures,
-  sampleBallotImages,
-} from '@votingworks/fixtures';
-import { join } from 'path';
-import { AddressInfo } from 'net';
-import fs from 'fs';
-import { execSync } from 'child_process';
-import { assert, deferred, ok, Result } from '@votingworks/basics';
-import {
-  buildMockInsertedSmartCardAuth,
-  InsertedSmartCardAuthApi,
-} from '@votingworks/auth';
-import { buildApp, Api } from '../../../../src/app';
-import {
-  createPrecinctScannerStateMachine,
-  Delays,
-} from '../../../../src/scanners/custom/state_machine';
+import { Api } from '../../../../src/app';
 import {
   createInterpreter,
   PrecinctScannerInterpreter,
 } from '../../../../src/interpret';
-import { createWorkspace, Workspace } from '../../../../src/util/workspace';
-import { Usb } from '../../../../src/util/usb';
 import {
-  PrecinctScannerState,
-  PrecinctScannerStatus,
-} from '../../../../src/types';
+  createPrecinctScannerStateMachine,
+  Delays,
+} from '../../../../src/scanners/custom/state_machine';
+import { createWorkspace, Workspace } from '../../../../src/util/workspace';
+import {
+  createApp,
+  createMockUsb,
+  expectStatus,
+  MockUsb,
+  waitForStatus,
+} from '../../app_helpers';
 
-type MockFileTree = MockFile | MockDirectory;
-type MockFile = Buffer;
-interface MockDirectory {
-  [name: string]: MockFileTree;
-}
-
-function writeMockFileTree(destinationPath: string, tree: MockFileTree): void {
-  if (Buffer.isBuffer(tree)) {
-    fs.writeFileSync(destinationPath, tree);
-  } else {
-    if (!fs.existsSync(destinationPath)) fs.mkdirSync(destinationPath);
-    for (const [name, child] of Object.entries(tree)) {
-      // Sleep 1ms to ensure that each file is created with a distinct timestamp
-      execSync('sleep 0.01');
-      writeMockFileTree(join(destinationPath, name), child);
-    }
-  }
-}
-
-interface MockUsb {
-  insertUsbDrive(contents: MockFileTree): void;
-  removeUsbDrive(): void;
-  mock: jest.Mocked<Usb>;
-}
-
-/**
- * Creates a mock of the Usb interface to USB drives. Simulates inserting and
- * removing a USB containing a tree of files and directories. Uses a temporary
- * directory on the filesystem to simulate the USB drive.
- */
-export function createMockUsb(): MockUsb {
-  let mockUsbTmpDir: tmp.DirResult | undefined;
-
-  const mock: jest.Mocked<Usb> = {
-    getUsbDrives: jest.fn().mockImplementation(() => {
-      if (mockUsbTmpDir) {
-        return Promise.resolve([
-          {
-            deviceName: 'mock-usb-drive',
-            mountPoint: mockUsbTmpDir.name,
-          },
-        ]);
-      }
-      return Promise.resolve([]);
-    }),
-  };
-
-  return {
-    mock,
-
-    insertUsbDrive(contents: MockFileTree) {
-      assert(!mockUsbTmpDir, 'Mock USB drive already inserted');
-      mockUsbTmpDir = tmp.dirSync({ unsafeCleanup: true });
-      writeMockFileTree(mockUsbTmpDir.name, contents);
-    },
-
-    removeUsbDrive() {
-      assert(mockUsbTmpDir, 'No mock USB drive to remove');
-      mockUsbTmpDir.removeCallback();
-      mockUsbTmpDir = undefined;
-    },
-  };
-}
-
-export async function expectStatus(
-  apiClient: grout.Client<Api>,
-  expectedStatus: {
-    state: PrecinctScannerState;
-  } & Partial<PrecinctScannerStatus>
-): Promise<void> {
-  const status = await apiClient.getScannerStatus();
-  expect(status).toEqual({
-    ballotsCounted: 0,
-    // TODO canUnconfigure should probably not be part of this endpoint - it's
-    // only needed on the admin screen
-    canUnconfigure: !expectedStatus?.ballotsCounted,
-    error: undefined,
-    interpretation: undefined,
-    ...expectedStatus,
-  });
-}
-
-export async function waitForStatus(
-  apiClient: grout.Client<Api>,
-  status: {
-    state: PrecinctScannerState;
-  } & Partial<PrecinctScannerStatus>
-): Promise<void> {
-  await waitForExpect(async () => {
-    await expectStatus(apiClient, status);
-  }, 1_000);
-}
-
-export async function createApp({
+export async function createCustomScannerApp({
   delays = {},
   mockCustomOptions = {},
   preconfiguredWorkspace,
@@ -184,24 +82,20 @@ export async function createApp({
     },
   });
   const mockUsb = createMockUsb();
-  const app = buildApp(
-    mockAuth,
-    precinctScannerMachine,
+
+  const { app, apiClient } = await createApp({
     interpreter,
-    workspace,
-    mockUsb.mock,
-    logger
-  );
-
-  const server = app.listen();
-  const { port } = server.address() as AddressInfo;
-  const baseUrl = `http://localhost:${port}/api`;
-
-  const apiClient = grout.createClient<Api>({ baseUrl });
+    logger,
+    mockAuth,
+    mockUsb,
+    precinctScannerMachine,
+    preconfiguredWorkspace: workspace,
+  });
 
   await expectStatus(apiClient, { state: 'connecting' });
   deferredConnect.resolve();
   await waitForStatus(apiClient, { state: 'no_paper' });
+
   return {
     apiClient,
     app,
@@ -241,58 +135,3 @@ export const ballotImages = {
     electionFamousNames2021Fixtures.machineMarkedBallotPage2.asFilePath(),
   ],
 } as const;
-
-// Loading of HMPB templates is slow, so in some tests we want to skip it by
-// removing the templates from the ballot package.
-export function createBallotPackageWithoutTemplates(
-  electionDefinition: ElectionDefinition
-): Buffer {
-  const dirPath = tmp.dirSync().name;
-  const zipPath = `${dirPath}.zip`;
-  fs.writeFileSync(
-    join(dirPath, 'election.json'),
-    electionDefinition.electionData
-  );
-  fs.writeFileSync(
-    join(dirPath, 'manifest.json'),
-    JSON.stringify({ ballots: [] })
-  );
-  execSync(`zip -j ${zipPath} ${dirPath}/*`);
-  return fs.readFileSync(zipPath);
-}
-const electionFamousNames2021WithoutTemplatesBallotPackageBuffer =
-  createBallotPackageWithoutTemplates(
-    electionFamousNames2021Fixtures.electionDefinition
-  );
-
-export async function configureApp(
-  apiClient: grout.Client<Api>,
-  mockUsb: MockUsb,
-  {
-    addTemplates = false,
-    precinctId,
-  }: {
-    addTemplates?: boolean;
-    precinctId?: PrecinctId;
-  } = {
-    addTemplates: false,
-  }
-): Promise<void> {
-  const ballotPackageBuffer = addTemplates
-    ? electionFamousNames2021Fixtures.ballotPackage.asBuffer()
-    : electionFamousNames2021WithoutTemplatesBallotPackageBuffer;
-  mockUsb.insertUsbDrive({
-    'ballot-packages': {
-      'test-ballot-package.zip': ballotPackageBuffer,
-    },
-  });
-  expect(await apiClient.configureFromBallotPackageOnUsbDrive()).toEqual(ok());
-
-  await apiClient.setPrecinctSelection({
-    precinctSelection: precinctId
-      ? singlePrecinctSelectionFor(precinctId)
-      : ALL_PRECINCTS_SELECTION,
-  });
-  await apiClient.setTestMode({ isTestMode: false });
-  await apiClient.setPollsState({ pollsState: 'polls_open' });
-}
